Revert "Kos Capacity in cfg"
This reverts commit 257bf8173170a3586a76a5fe044eb79aa5b07b76.

+++ Persistance/Harddisk.cs
@@ -6,9 +6,9 @@ namespace kOS.Persistance
 {
     public class Harddisk : Volume
     {
-        public Harddisk(float size)
+        public Harddisk(int size)
         {
-            Capacity = (int)size;
+            Capacity = size;
         }
 
         public Harddisk(ConfigNode node)
